rotation almost done

+++ rgb.c
@@ -1,30 +1,30 @@
 #include "fdf.h"
 
-static int	get_red(float z, t_fdf *s)
+static int	get_red(double z, t_fdf *s)
 {
-	float	persent;
+	double	persent;
 
 	if (!s)
 		free_exit(s, "get_red - empty pointer found");
 	if (z > 0 && z > s->max.z / 5)
 	{
 		persent = (z - s->max.z / 5) / (s->max.z) + 0.2;
-		return ((int)((float)255 * persent));
+		return ((int)((double)255 * persent));
 	}
 	return (0);
 }
 
-static int	get_green(float z, t_fdf *s)
+static int	get_green(double z, t_fdf *s)
 {
-	float	persent;
-	float	min_z;
+	double	persent;
+	double	min_z;
 
 	if (!s)
 		free_exit(s, "get_green - empty pointer found");
 	if (z > 0 && z < s->max.z * 4 / 5)
 	{
 		persent = z / (s->max.z * 4 / 5);
-		return ((int)((float)255 * (1 - persent)));
+		return ((int)((double)255 * (1 - persent)));
 	}
 	if (z == 0)
 			return (255);
@@ -33,15 +33,15 @@ static int	get_green(float z, t_fdf *s)
 		min_z = - s->min.z;
 		z = -z;
 		persent = 1 - z / (min_z / 3);
-		return ((int)((float)255 * persent));
+		return ((int)((double)255 * persent));
 	}
 	return (0);
 }
 
-static int	get_blue(float z, t_fdf *s)
+static int	get_blue(double z, t_fdf *s)
 {
-	float	persent;
-	float	min_z;
+	double	persent;
+	double	min_z;
 
 	if (!s)
 		free_exit(s, "get_blue - empty pointer found");
@@ -50,12 +50,12 @@ static int	get_blue(float z, t_fdf *s)
 		min_z = -(s->min.z);
 		z = -z;
 		persent = z / min_z;
-		return ((int)((float)255 * persent));
+		return ((int)((double)255 * persent));
 	}
 	return (0);
 }
 
-int			get_point_color(float z, t_fdf *s)
+int			get_point_color(double z, t_fdf *s)
 {
 	int		r;
 	int		g;
